# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a campus app that shows which students are in class right now.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type Campus = typeof Campus[keyof typeof Campus];


export const Campus = {
  hyderabad: 'hyderabad',
  mohali: 'mohali',
} as const;

export interface StudentSummary {
  id: number;
  name: string;
  email: string;
  campus: Campus;
  section: string;
}

export type Student = StudentSummary & {
  studyGroup: string;
};

export interface ClassSession {
  courseCode: string;
  courseName: string;
  section: string;
  date: string;
  /** 24-hour HH:MM in Asia/Kolkata time */
  startTime: string;
  /** 24-hour HH:MM in Asia/Kolkata time */
  endTime: string;
  /** @nullable */
  room: string | null;
}

export interface StudentStatus {
  student: StudentSummary;
  isInClass: boolean;
  currentSession: ClassSession | null;
  /** The next session later today, if any */
  nextSession: ClassSession | null;
}

export interface FriendEntry {
  student: StudentSummary;
  isInClass: boolean;
  currentSession: ClassSession | null;
  nextSession: ClassSession | null;
}

export interface FriendInput {
  friendId: number;
}

export type ListStudentsParams = {
campus?: Campus;
/**
 * Case-insensitive match against student name or email
 */
search?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};