# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for a campus app that shows which students are in class right now.
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: students
    description: Student roster, schedules, and live class status
  - name: friends
    description: Managing a student's friends list
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: Search the student roster
      description: Returns students matching an optional campus filter and name/email search term. Used for onboarding (picking your own name) and for searching/adding friends.
      parameters:
        - name: campus
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Campus"
        - name: search
          in: query
          required: false
          description: Case-insensitive match against student name or email
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Matching students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentSummary"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a single student's profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /students/{id}/status:
    get:
      operationId: getStudentStatus
      tags: [students]
      summary: Get a student's live class status right now
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Live status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentStatus"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /students/{id}/schedule:
    get:
      operationId: getStudentSchedule
      tags: [students]
      summary: Get a student's class schedule for today
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Today's schedule, ordered by start time
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassSession"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /students/{id}/schedule/{date}:
    get:
      operationId: getStudentScheduleForDate
      tags: [students]
      summary: Get a student's class schedule for a specific day
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: path
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}-\\d{2}$"
      responses:
        "200":
          description: Schedule for the given day, ordered by start time
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassSession"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /students/{id}/friends:
    get:
      operationId: listFriends
      tags: [friends]
      summary: List a student's friends with live status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Friends with live class status, free friends first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendEntry"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: addFriend
      tags: [friends]
      summary: Add a friend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendInput"
      responses:
        "201":
          description: Friend added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendEntry"
        "400":
          description: Invalid request (e.g. adding yourself, unknown student, or already friends)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /students/{id}/friends/{friendId}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: friendId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Friend removed
        "404":
          description: Friendship not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Campus:
      type: string
      enum: [hyderabad, mohali]
    StudentSummary:
      type: object
      required: [id, name, email, campus, section]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        campus:
          $ref: "#/components/schemas/Campus"
        section:
          type: string
    Student:
      allOf:
        - $ref: "#/components/schemas/StudentSummary"
        - type: object
          required: [studyGroup]
          properties:
            studyGroup:
              type: string
    ClassSession:
      type: object
      required: [courseCode, courseName, section, date, startTime, endTime, room]
      properties:
        courseCode:
          type: string
        courseName:
          type: string
        section:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
          description: 24-hour HH:MM in Asia/Kolkata time
        endTime:
          type: string
          description: 24-hour HH:MM in Asia/Kolkata time
        room:
          type: ["string", "null"]
    StudentStatus:
      type: object
      required: [student, isInClass, currentSession, nextSession]
      properties:
        student:
          $ref: "#/components/schemas/StudentSummary"
        isInClass:
          type: boolean
        currentSession:
          oneOf:
            - $ref: "#/components/schemas/ClassSession"
            - type: "null"
        nextSession:
          description: The next session later today, if any
          oneOf:
            - $ref: "#/components/schemas/ClassSession"
            - type: "null"
    FriendEntry:
      type: object
      required: [student, isInClass, currentSession, nextSession]
      properties:
        student:
          $ref: "#/components/schemas/StudentSummary"
        isInClass:
          type: boolean
        currentSession:
          oneOf:
            - $ref: "#/components/schemas/ClassSession"
            - type: "null"
        nextSession:
          oneOf:
            - $ref: "#/components/schemas/ClassSession"
            - type: "null"
    FriendInput:
      type: object
      required: [friendId]
      properties:
        friendId:
          type: integer